pyproject.toml
[project]
name = "townwars-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["pygame-ce"]

[tool.ruff.lint]
select = ["ALL"]

ignore = [
  "D100",
  "D101",
  "D102",
  "D103",
  "D104",
  "D107",
  "D211",
  "EM101",
  "G004",
  "TD003",
  "TRY003",
  ]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402"]

[tool.ruff.format]
quote-style = "double"